Updated timer functionality in index.js
- updated the initial display of the timer on lines 25 and 26

- created new variable "setCounter" on line 28 to keep track of the number of sets completed by the user

- Updated the "startTimer" function.

seconds decrease by 1 each second. If the seconds are lower than 0, then seconds is set to 59 and minutes decrease by 1 to reduce the timer until it eventually reached 00:00.

created variables "currentMinutes" on line 42 and "currentSeconds" on line 43. If the minutes or seconds are lower than 10, then a 0 is added before the number to keep up the appearances of a normal timer. The timers text content is then set to these values on line 44.

if the minutes and seconds are both equal to 0, then the timer stops and the set counter increases by one.

The function runs again each second

- removed the "el." before the start, increment and decrement buttons

- removed the "prepareHandles" call at the bottom of the file.